+++ packages/atproto_core/lib/src/base_service.dart
@@ -17,7 +17,7 @@ abstract class _Service {
     final UserContext userContext = UserContext.authRequired,
     final Map<String, dynamic>? parameters,
     required final xrpc.To<T> to,
-    final xrpc.JsonConverter? converter,
+    final xrpc.ResponseAdaptor? adaptor,
   });
 
   Future<xrpc.XRPCResponse<T>> post<T>(
@@ -44,8 +44,7 @@ abstract class _Service {
     final String? service,
     final Map<String, dynamic>? parameters,
     final xrpc.To<T>? to,
-    final xrpc.Decoder? decoder,
-    final xrpc.JsonConverter? converter,
+    final xrpc.ResponseAdaptor? adaptor,
   });
 }
 
@@ -86,7 +85,7 @@ abstract class BaseService implements _Service {
     final UserContext userContext = UserContext.authRequired,
     final Map<String, dynamic>? parameters,
     required final xrpc.To<T> to,
-    final xrpc.JsonConverter? converter,
+    final xrpc.ResponseAdaptor? adaptor,
   }) async =>
       await _context.get(
         xrpc.NSID.create(
@@ -98,7 +97,7 @@ abstract class BaseService implements _Service {
         service: _service,
         parameters: parameters,
         to: to,
-        converter: converter,
+        adaptor: adaptor,
         getClient: _mockedGetClient,
       );
 
@@ -153,8 +152,7 @@ abstract class BaseService implements _Service {
     final String? service,
     final Map<String, dynamic>? parameters,
     final xrpc.To<T>? to,
-    final xrpc.Decoder? decoder,
-    final xrpc.JsonConverter? converter,
+    final xrpc.ResponseAdaptor? adaptor,
   }) async =>
       await _context.stream(
         methodId,
@@ -162,8 +160,7 @@ abstract class BaseService implements _Service {
         service: service,
         parameters: parameters,
         to: to,
-        decoder: decoder,
-        converter: converter,
+        adaptor: adaptor,
       );
 
   /// Returns the NSID based on this service and [methodName].